+++ whiplash/src/parser/grammar/lr1_item/item.rs
@@ -1,4 +1,4 @@
-use crate::parser::lr0_item;
+use crate::parser::grammar::lr0_item;
 use super::look_ahead;
 use std::fmt;
