+++ whiplash/src/parser/mod.rs
@@ -1,3 +1,2 @@
 pub mod node;
-pub mod lr0_item;
-pub mod lr1_item;
+pub mod grammar;
